implementando os métodos do agendamento service

+++ Agenda.Api/Services/Agendamentos/AgendamentoService.cs
@@ -1,5 +1,7 @@
 ﻿using Agenda.Api.Data.Repository;
+using Agenda.Api.Infra;
 using Agenda.Api.Models;
+using Agenda.Api.Services.Medicos;
 using System.Collections.Generic;
 
 namespace Agenda.Api.Services.Agendamentos
@@ -7,37 +9,62 @@ namespace Agenda.Api.Services.Agendamentos
     public class AgendamentoService : IAgendamentoService
     {
         private readonly IAgendamentoRepository _agendamentoRepository;
+        private readonly Notification _notification;
+        private readonly IMedicoRepository _medicoRepository;
 
-        public AgendamentoService(AgendamentoRepository agendamentoRepository)
+        public AgendamentoService(AgendamentoRepository agendamentoRepository, Notification notification, IMedicoRepository medicoRepository)
         {
             _agendamentoRepository = agendamentoRepository;
+            _notification = notification;
+            _medicoRepository = medicoRepository;
         }
 
         public Agendamento Delete(int id)
         {
+            var agendamento = _agendamentoRepository.Get(id);
+
+            if (agendamento == null)
+            {
+                _notification.Add("Agendamento não encontrado");
+                return null;
+            }
+
             return _agendamentoRepository.Delete(id);
         }
 
-        public IEnumerable<Agendamento> Get(int? idMedico)
-        {
-            return _agendamentoRepository.Get(idMedico);
+        public IEnumerable<Agendamento> Get(int? idMedico) => _agendamentoRepository.Get(idMedico);
 
+        public Agendamento Get(int id) => _agendamentoRepository.Get(id);
+
+        public Agendamento Post(Agendamento agendamento)
+        {
+            if (!IsValid(agendamento))
+                return agendamento;
+
+            return _agendamentoRepository.Post(agendamento);
         }
 
-        public Agendamento Get(int id)
+        public void Put(Agendamento agendamento)
         {
-            return _agendamentoRepository.Get(id);
+            if (!IsValid(agendamento))
+                return;
 
+            _agendamentoRepository.Put(agendamento);
         }
 
-        public Agendamento Post(Agendamento gendamento)
+        private bool IsValid(Agendamento agendamento)
         {
-            return _agendamentoRepository.Post(gendamento);
-        }
+            // validando MÉDICO
+            var medico = _medicoRepository.Get(agendamento.Medico.Id);
+            if (medico == null)
+                return _notification.AddWithReturn<bool>("Médico não encontrado");
 
-        public void Put(Agendamento gendamento)
-        {
-            _agendamentoRepository.Put(gendamento);
+            // validando PACIENTE
+            var paciente = _medicoRepository.Get(agendamento.Paciente.Id);
+            if (paciente == null)
+                return _notification.AddWithReturn<bool>("Paciente não encontrado");
+
+            return true;
         }
     }
 }
